requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1.3
django-recurrence==1.7.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
